# Add YOLOv6 quantization and auto compression example