--- a/tools/auto_compression/configs/yolov6_tiny_qat_dis.yaml
+++ b/tools/auto_compression/configs/yolov6_tiny_qat_dis.yaml
@@ -1,0 +1,32 @@
+Global:
+  model_dir: ./yolov6t.onnx
+  image_path: None   # If image_path is set, it will be trained directly based on unlabeled images, no need to set the COCO dataset path.
+  coco_dataset_dir: /paddle/dataset/coco/
+  coco_train_image_dir: train2017
+  coco_train_anno_path: annotations/instances_train2017.json
+  coco_val_image_dir: val2017
+  coco_val_anno_path: annotations/instances_val2017.json
+  arch: YOLOv6
+
+Distillation:
+  alpha: 1.0
+  loss: soft_label
+
+Quantization:
+  onnx_format: true
+  activation_quantize_type: 'moving_average_abs_max'
+  quantize_op_types:
+  - conv2d
+  - depthwise_conv2d
+
+TrainConfig:
+  train_iter: 8000
+  eval_iter: 1000
+  learning_rate:
+    type: CosineAnnealingDecay
+    learning_rate: 0.00003
+    T_max: 8000
+  optimizer_builder:
+    optimizer:
+      type: SGD
+    weight_decay: 0.00004