We used [PaddleSlim](https://github.com/PaddlePaddle/PaddleSlim) ACT(Auto Compression Toolkit) to quantize and compress YOLOv6, and the results on T4  with TensorRT are as follows:

| Model | Base mAP<sup>val<br>0.5:0.95  | Quant mAP<sup>val<br>0.5:0.95 | Latency<sup><small>FP32</small><sup><br><sup> | Latency<sup><small>FP16</small><sup><br><sup> | Latency<sup><small>INT8</small><sup><br><sup> | 
| :-------- |:-------- |:--------: | :--------: | :---------------------: | :----------------: | 
| YOLOv6s |  42.4   | 41.3  |  9.06ms  |   2.90ms   |  **1.83ms**  | 
| YOLOv6-Tiny  |  40.8   | 39.0 |  5.06ms  |   2.32ms   |  **1.68ms** |